refactor(wav2vec): rename wav2vec submodule to huggingface
